A couple more reorganization items
