+++ forecasting_tools/__init__.py
@@ -12,7 +12,7 @@ from forecasting_tools.agents_and_tools.base_rates.niche_list_researcher import 
 from forecasting_tools.agents_and_tools.base_rates.niche_list_researcher import (
     NicheListResearcher as NicheListResearcher,
 )
-from forecasting_tools.agents_and_tools.question_generators.question_generator import (
+from forecasting_tools.agents_and_tools.deprecated.question_generator import (
     QuestionGenerator as QuestionGenerator,
 )
 from forecasting_tools.agents_and_tools.question_generators.topic_generator import (
