+++ forecasting_tools/agents_and_tools/question_generators/question_decomposer.py
@@ -7,7 +7,7 @@ from typing import Literal
 
 from pydantic import BaseModel
 
-from forecasting_tools.agents_and_tools.misc_tools import (
+from forecasting_tools.agents_and_tools.minor_tools import (
     perplexity_quick_search,
     query_asknews,
 )
