+++ forecasting_tools/agents_and_tools/research/find_a_dataset.py
@@ -3,7 +3,9 @@ import logging
 
 from pydantic import BaseModel
 
-from forecasting_tools.agents_and_tools.misc_tools import perplexity_pro_search
+from forecasting_tools.agents_and_tools.minor_tools import (
+    perplexity_pro_search,
+)
 from forecasting_tools.agents_and_tools.other.data_analyzer import DataAnalyzer
 from forecasting_tools.agents_and_tools.research.computer_use import (
     ComputerUse,
