+++ scripts/run_optimizer.py
@@ -1,7 +1,7 @@
 import asyncio
 import logging
 
-from forecasting_tools.agents_and_tools.misc_tools import (
+from forecasting_tools.agents_and_tools.minor_tools import (
     perplexity_quick_search_low_context,
     query_asknews,
 )
